tweetrend-realtime-api-server/src/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
blinker==1.4
cassandra-driver==3.24.0
cbor==1.0.0
certifi==2018.1.18
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
cloud-init==20.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.3.1
Cython==0.26.1
daphne==3.0.1
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
ec2-hibinit-agent==1.0.0
filelock==3.0.12
geomet==0.2.1.post1
hibagent==1.0.1
hiredis==1.1.0
httplib2==0.9.2
hyperlink==21.0.0
idna==3.1
importlib-metadata==3.4.0
importlib-resources==5.0.0
incremental==17.5.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lz4==0.10.1
MarkupSafe==1.0
msgpack==1.0.2
netifaces==0.10.4
oauthlib==2.0.6
olefile==0.45.1
PAM==0.4.2
Pillow==5.1.0
py-ubjson==0.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.5.3
PyNaCl==1.1.2
pyOpenSSL==20.0.1
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
python-snappy==0.5
PyTrie==0.2
pytz==2020.5
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==20.3.0
wsaccel==0.6.2
zipp==3.4.0
zope.interface==5.2.0
